+++ src/app/(main)/cart/order/page.tsx
@@ -3,7 +3,6 @@ import OrderPageContent from "./_components/OrderPageContent";
 
 export default async function OrderPage({ searchParams }: { searchParams: Promise<{ cartItemId?: string }> }) {
   const params = await searchParams;
-  const cartItemId = params.cartItemId;
 
-  return <OrderPageContent cartItemId={cartItemId} />;
+  return <OrderPageContent cartItemId={params.cartItemId} />;
 }
